backend/requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-toolbelt==0.0.1
djangorestframework==3.6.4
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
static3==0.7.0
